feat: Allow Users to send and receive many messages continuously
Allow a User to send and receive messages without having to go back to
the main CLI menu.

Change the Pub/Sub message broker back to NATS due to some weird and
hard to fix errors with Apache Pulsar.
Change the .env configuration file to the connection params for NATS.